Added a helper method: calculateNextOddNumber()
Since it is not very clear why currentOddNumber get incremented by 2 in the do-while loop in the calculatePrimes() method, a helper method calculateNextOddNumber() was added. 
This method makes it easier to understand why currentOddNumber would need to be incremented by 2. Also, by adding the method, we make the computational body of the finding the prime number easier to understand (i.e. without having to figure out what the purpose of each instruction is)